+++ server/src/main/java/com/seniclass/server/domain/student/controller/VideoProgressController.java
@@ -2,8 +2,8 @@ package com.seniclass.server.domain.student.controller;
 
 import com.seniclass.server.domain.auth.domain.RequireAuth;
 import com.seniclass.server.domain.auth.enums.UserRole;
-import com.seniclass.server.domain.student.dto.VideoProgressRequest;
-import com.seniclass.server.domain.student.dto.VideoProgressResponse;
+import com.seniclass.server.domain.student.dto.request.VideoProgressRequest;
+import com.seniclass.server.domain.student.dto.response.VideoProgressResponse;
 import com.seniclass.server.domain.student.service.VideoProgressService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
