+++ server/src/main/java/com/seniclass/server/domain/student/dto/request/LectureQnaRequest.java
@@ -1,4 +1,4 @@
-package com.seniclass.server.domain.student.dto;
+package com.seniclass.server.domain.student.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
