+++ server/src/main/java/com/seniclass/server/domain/student/service/AssignmentSubmissionServiceImpl.java
@@ -6,8 +6,8 @@ import com.seniclass.server.domain.lecture.domain.Assignment;
 import com.seniclass.server.domain.lecture.repository.AssignmentRepository;
 import com.seniclass.server.domain.student.domain.AssignmentSubmission;
 import com.seniclass.server.domain.student.domain.Student;
-import com.seniclass.server.domain.student.dto.AssignmentSubmissionFileRequest;
-import com.seniclass.server.domain.student.dto.AssignmentSubmissionResponse;
+import com.seniclass.server.domain.student.dto.request.AssignmentSubmissionFileRequest;
+import com.seniclass.server.domain.student.dto.response.AssignmentSubmissionResponse;
 import com.seniclass.server.domain.student.exception.errorcode.AssignmentSubmissionErrorCode;
 import com.seniclass.server.domain.student.exception.errorcode.StudentErrorCode;
 import com.seniclass.server.domain.student.repository.AssignmentSubmissionRepository;
