+++ server/src/main/java/com/seniclass/server/domain/student/service/LectureBookmarkServiceImpl.java
@@ -5,8 +5,8 @@ import com.seniclass.server.domain.lecture.domain.Lecture;
 import com.seniclass.server.domain.lecture.repository.LectureRepository;
 import com.seniclass.server.domain.student.domain.LectureBookmark;
 import com.seniclass.server.domain.student.domain.Student;
-import com.seniclass.server.domain.student.dto.LectureBookmarkRequest;
-import com.seniclass.server.domain.student.dto.LectureBookmarkResponse;
+import com.seniclass.server.domain.student.dto.request.LectureBookmarkRequest;
+import com.seniclass.server.domain.student.dto.response.LectureBookmarkResponse;
 import com.seniclass.server.domain.student.exception.errorcode.LectureBookmarkErrorCode;
 import com.seniclass.server.domain.student.exception.errorcode.StudentErrorCode;
 import com.seniclass.server.domain.student.repository.LectureBookmarkRepository;
