+++ server/src/main/java/com/seniclass/server/domain/student/service/LectureEnrollmentServiceImpl.java
@@ -5,8 +5,8 @@ import com.seniclass.server.domain.lecture.domain.Lecture;
 import com.seniclass.server.domain.lecture.repository.LectureRepository;
 import com.seniclass.server.domain.student.domain.LectureEnrollment;
 import com.seniclass.server.domain.student.domain.Student;
-import com.seniclass.server.domain.student.dto.LectureEnrollmentRequest;
-import com.seniclass.server.domain.student.dto.LectureEnrollmentResponse;
+import com.seniclass.server.domain.student.dto.request.LectureEnrollmentRequest;
+import com.seniclass.server.domain.student.dto.response.LectureEnrollmentResponse;
 import com.seniclass.server.domain.student.exception.errorcode.LectureEnrollmentErrorCode;
 import com.seniclass.server.domain.student.exception.errorcode.StudentErrorCode;
 import com.seniclass.server.domain.student.repository.LectureEnrollmentRepository;
