+++ server/src/main/java/com/seniclass/server/domain/student/service/LectureQnaServiceImpl.java
@@ -5,9 +5,9 @@ import com.seniclass.server.domain.lecture.domain.Lecture;
 import com.seniclass.server.domain.lecture.repository.LectureRepository;
 import com.seniclass.server.domain.student.domain.LectureQna;
 import com.seniclass.server.domain.student.domain.Student;
-import com.seniclass.server.domain.student.dto.LectureQnaAnswerRequest;
-import com.seniclass.server.domain.student.dto.LectureQnaRequest;
-import com.seniclass.server.domain.student.dto.LectureQnaResponse;
+import com.seniclass.server.domain.student.dto.request.LectureQnaAnswerRequest;
+import com.seniclass.server.domain.student.dto.request.LectureQnaRequest;
+import com.seniclass.server.domain.student.dto.response.LectureQnaResponse;
 import com.seniclass.server.domain.student.exception.errorcode.LectureQnaErrorCode;
 import com.seniclass.server.domain.student.exception.errorcode.StudentErrorCode;
 import com.seniclass.server.domain.student.repository.LectureQnaRepository;
