+++ server/src/main/java/com/seniclass/server/domain/student/service/VideoProgressService.java
@@ -1,7 +1,7 @@
 package com.seniclass.server.domain.student.service;
 
-import com.seniclass.server.domain.student.dto.VideoProgressRequest;
-import com.seniclass.server.domain.student.dto.VideoProgressResponse;
+import com.seniclass.server.domain.student.dto.request.VideoProgressRequest;
+import com.seniclass.server.domain.student.dto.response.VideoProgressResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
