+++ server/src/main/java/com/seniclass/server/domain/student/service/VideoProgressServiceImpl.java
@@ -5,8 +5,8 @@ import com.seniclass.server.domain.lecture.domain.Video;
 import com.seniclass.server.domain.lecture.repository.VideoRepository;
 import com.seniclass.server.domain.student.domain.Student;
 import com.seniclass.server.domain.student.domain.VideoProgress;
-import com.seniclass.server.domain.student.dto.VideoProgressRequest;
-import com.seniclass.server.domain.student.dto.VideoProgressResponse;
+import com.seniclass.server.domain.student.dto.request.VideoProgressRequest;
+import com.seniclass.server.domain.student.dto.response.VideoProgressResponse;
 import com.seniclass.server.domain.student.exception.errorcode.StudentErrorCode;
 import com.seniclass.server.domain.student.exception.errorcode.VideoProgressErrorCode;
 import com.seniclass.server.domain.student.repository.StudentRepository;
