+++ server/src/test/java/com/seniclass/server/domain/student/service/AssignmentSubmissionServiceTest.java
@@ -10,8 +10,8 @@ import com.seniclass.server.domain.lecture.domain.Lecture;
 import com.seniclass.server.domain.lecture.repository.AssignmentRepository;
 import com.seniclass.server.domain.student.domain.AssignmentSubmission;
 import com.seniclass.server.domain.student.domain.Student;
-import com.seniclass.server.domain.student.dto.AssignmentSubmissionFileRequest;
-import com.seniclass.server.domain.student.dto.AssignmentSubmissionResponse;
+import com.seniclass.server.domain.student.dto.request.AssignmentSubmissionFileRequest;
+import com.seniclass.server.domain.student.dto.response.AssignmentSubmissionResponse;
 import com.seniclass.server.domain.student.repository.AssignmentSubmissionRepository;
 import com.seniclass.server.domain.student.repository.StudentRepository;
 import com.seniclass.server.global.exception.CommonException;
